feat: reduce bundle size from wso2 deployment handler
