+++ lib/src/wso2/utils.ts
@@ -1,89 +1,7 @@
 /* eslint-disable no-console */
-import { existsSync } from 'fs';
-
-import { Duration, ScopedAws } from 'aws-cdk-lib/core';
-import { PolicyStatement } from 'aws-cdk-lib/aws-iam';
-import { Runtime } from 'aws-cdk-lib/aws-lambda';
-import { RetentionDays } from 'aws-cdk-lib/aws-logs';
-import { Provider } from 'aws-cdk-lib/custom-resources';
-import { Construct } from 'constructs';
 import { GetSecretValueCommand, SecretsManagerClient } from '@aws-sdk/client-secrets-manager';
 
-import { BaseNodeJsFunction } from '../lambda/lambda-base';
-import { EventType } from '../lambda/types';
-
-import { RetryOptions, Wso2BaseProperties } from './types';
-
-export const addLambdaAndProviderForWso2Operations = (args: {
-  scope: Construct;
-  id: string;
-  props: Wso2BaseProperties;
-  baseDir: string;
-}): { customResourceProvider: Provider; customResourceFunction: BaseNodeJsFunction } => {
-  const logGroupRetention =
-    args.props.customResourceConfig?.logGroupRetention ?? RetentionDays.ONE_MONTH;
-
-  const { accountId, region } = new ScopedAws(args.scope);
-
-  // resolve the entry file from workspace (.ts file), or
-  // from the dist dir (.js file) when being used as a lib
-  let wso2LambdaEntry = `${args.baseDir}/handler/index.ts`;
-  if (!existsSync(wso2LambdaEntry)) {
-    wso2LambdaEntry = `${args.baseDir}/handler/index.js`;
-  }
-
-  // lambda function used for invoking WSO2 APIs during CFN operations
-  const customResourceFunction = new BaseNodeJsFunction(args.scope, `${args.id}-custom-lambda`, {
-    ...args.props.customResourceConfig,
-    stage: 'wso2-custom-lambda',
-    timeout: Duration.minutes(10),
-    memorySize: 256,
-    runtime: Runtime.NODEJS_18_X,
-    eventType: EventType.CustomResource,
-    createLiveAlias: false,
-    createDefaultLogGroup: true,
-    entry: wso2LambdaEntry,
-    initialPolicy: [
-      PolicyStatement.fromJson({
-        Effect: 'Allow',
-        Action: 'secretsmanager:GetSecretValue',
-        Resource: `arn:aws:secretsmanager:${region}:${accountId}:secret:${args.props.wso2Config.credentialsSecretId}*`,
-      }),
-    ],
-    logGroupRetention,
-    // allow all outbound by default
-    allowAllOutbound: typeof args.props.customResourceConfig?.network !== 'undefined',
-  });
-
-  if (args.props.customResourceConfig?.logGroupRemovalPolicy) {
-    customResourceFunction.nodeJsFunction.applyRemovalPolicy(
-      args.props.customResourceConfig.logGroupRemovalPolicy,
-    );
-  }
-
-  const customResourceProvider = new Provider(args.scope, `${args.id}-custom-provider`, {
-    onEventHandler: customResourceFunction.nodeJsFunction,
-  });
-
-  return { customResourceProvider, customResourceFunction };
-};
-
-export const getSecretValue = async (secretId: string): Promise<string> => {
-  const client = new SecretsManagerClient();
-  const response = await client.send(
-    new GetSecretValueCommand({
-      SecretId: secretId,
-    }),
-  );
-  if (response.SecretString) {
-    return response.SecretString;
-  }
-  if (!response.SecretBinary) {
-    throw new Error('Invalid type of secret found');
-  }
-  const buff = Buffer.from(response.SecretBinary);
-  return buff.toString('ascii');
-};
+import { RetryOptions } from './types';
 
 const defaultRetryOpts = {
   checkRetries: {
@@ -103,6 +21,7 @@ const defaultRetryOpts = {
     // 2000, 3000
   },
 };
+
 export const applyRetryDefaults = (retryOptions?: RetryOptions): RetryOptions => {
   const ropts: RetryOptions = {
     // default config for backoff
@@ -137,4 +56,21 @@ export const applyRetryDefaults = (retryOptions?: RetryOptions): RetryOptions =>
 
 export const truncateStr = (str: string, size: number): string => {
   return str.substring(0, Math.min(str.length, size));
+};
+
+export const getSecretValue = async (secretId: string): Promise<string> => {
+  const client = new SecretsManagerClient();
+  const response = await client.send(
+    new GetSecretValueCommand({
+      SecretId: secretId,
+    }),
+  );
+  if (response.SecretString) {
+    return response.SecretString;
+  }
+  if (!response.SecretBinary) {
+    throw new Error('Invalid type of secret found');
+  }
+  const buff = Buffer.from(response.SecretBinary);
+  return buff.toString('ascii');
 };
